Fuel system added with ending

+++ System/GameData.cs
@@ -9,6 +9,6 @@ public static class GameData
     public static int PackageNum {get;set;}
     public static string Base {get;set;}
     public static string Door {get;set;}
-    public static int Fuel {get;set;}
+    public static float Fuel {get;set;}
     public static bool InSpace {get;set;}
 }
